+++ schemacrawler-ermodel/src/test/java/schemacrawler/ermodel/weakassociations/IdMatcherTest.java
@@ -21,7 +21,7 @@ public class IdMatcherTest {
     final Column pkColumn = mockColumn(pkTable, "order_id", true);
 
     final IdMatcher matcher = new IdMatcher();
-    assertThat(matcher.test(new WeakAssociationColumnReference(fkColumn, pkColumn)), is(true));
+    assertThat(matcher.test(new WeakColumnReference(fkColumn, pkColumn)), is(true));
   }
 
   @Test
@@ -33,7 +33,7 @@ public class IdMatcherTest {
     final Column pkColumn = mockColumn(pkTable, "id", true);
 
     final IdMatcher matcher = new IdMatcher();
-    assertThat(matcher.test(new WeakAssociationColumnReference(fkColumn, pkColumn)), is(true));
+    assertThat(matcher.test(new WeakColumnReference(fkColumn, pkColumn)), is(true));
   }
 
   @Test
@@ -45,7 +45,7 @@ public class IdMatcherTest {
     final Column pkColumn = mockColumn(pkTable, "order_id", true);
 
     final IdMatcher matcher = new IdMatcher();
-    assertThat(matcher.test(new WeakAssociationColumnReference(fkColumn, pkColumn)), is(true));
+    assertThat(matcher.test(new WeakColumnReference(fkColumn, pkColumn)), is(true));
   }
 
   @Test
@@ -57,7 +57,7 @@ public class IdMatcherTest {
     final Column pkColumn = mockColumn(pkTable, "id", true);
 
     final IdMatcher matcher = new IdMatcher();
-    assertThat(matcher.test(new WeakAssociationColumnReference(fkColumn, pkColumn)), is(true));
+    assertThat(matcher.test(new WeakColumnReference(fkColumn, pkColumn)), is(true));
   }
 
   @Test
@@ -69,7 +69,7 @@ public class IdMatcherTest {
     final Column pkColumn = mockColumn(pkTable, "id", true);
 
     final IdMatcher matcher = new IdMatcher();
-    assertThat(matcher.test(new WeakAssociationColumnReference(fkColumn, pkColumn)), is(false));
+    assertThat(matcher.test(new WeakColumnReference(fkColumn, pkColumn)), is(false));
   }
 
   @Test
@@ -81,7 +81,7 @@ public class IdMatcherTest {
     final Column pkColumn = mockColumn(pkTable, "order", true);
 
     final IdMatcher matcher = new IdMatcher();
-    assertThat(matcher.test(new WeakAssociationColumnReference(fkColumn, pkColumn)), is(false));
+    assertThat(matcher.test(new WeakColumnReference(fkColumn, pkColumn)), is(false));
   }
 
   @Test
@@ -96,7 +96,7 @@ public class IdMatcherTest {
     final Column pkColumn = mockColumn(pkTable, "order_id", true);
 
     final IdMatcher matcher = new IdMatcher();
-    assertThat(matcher.test(new WeakAssociationColumnReference(fkColumn, pkColumn)), is(false));
+    assertThat(matcher.test(new WeakColumnReference(fkColumn, pkColumn)), is(false));
   }
 
   private Column mockColumn(final Table parent, final String name, final boolean partOfPrimaryKey) {