+++ changelog.md
@@ -5,6 +5,7 @@ SchemaCrawler Core release notes.
 <a name="v17.5.1"></a>
 ## Release 17.5.1 - 2026-01-25
 - Reduce number of weak association matches
+- Add weak associations to `ERModel`
 
 
 <a name="v17.5.0"></a>